Move legacy types to directory not built to production
