+++ packages/audiences-react/src/AudienceSelector/legacy-types.ts
@@ -1,0 +1,18 @@
+type Option = {
+  id: string
+  name: string
+}
+
+export type Filter = {
+  id: string
+  size: number // the number of people that matches this filter audience
+  userCount: number
+  departments?: Option[]
+  titles?: Option[]
+  territories?: Option[]
+}
+
+export type Member = {
+  id: string
+  name: string
+}
